requirements.txt
keyboard==0.13.5
mouse==0.7.1
mss==6.1.0
numpy==1.23.0
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
pytesseract==0.3.9
python-Levenshtein==0.12.2
thefuzz==0.19.0
